Hide tags/branches in RTD theme. More themes.
It's possible for only tags or (exclusive) branches to be in Versions.
Read the docs theme should hide either when they're empty. Updating
versions.html for this.

Adding __bool__/__nonzero__ methods to Versions to support if statements
in HTML templates. Also returning lists instead of yielding remotes so
.tags and .branches can be in if statements.

Adding "id" key to Versions.remotes list items. One SHA can be used by
multiple tags/branches, a tag and a branch can share the same name, it's
technically possible for two refs to have the same date, and URL is not
always set. type/name is what ls-remote uses and should be unique.

Testing classic and sphinxdoc themes.