Beginning to roll back mistake.
SOURCE was never meant to be a currently existing file path. It's
supposed to be a relative path to one of many git refs that may not
currently be pulled. Starting to roll back.

Adding single-letter arg names for all arguments.

Filtering out pull requests in RE_REMOTE regex. Only tags and heads
(branches) will be looked up.