Semi-ignore .doctrees and searchindex.js. --depth
Soon main_push() will be defined and it will call main_build. Need to
have it specify a different detination path. Adding detination argument
to main_build() for this.

Cloning with --depth=1 in clone().

Not committing in commit_and_push() if the only files that have changed
are those Sphinx generated files that always change even when source RST
files do not.