Handing git commit errors. Commit message file.
Added PyPI credentials for future release.

Handling error on git commit. Forgot that on Travis or any new
environment if the user forgets to specify a name and email, git commit
will error out asking for that info.

Moving commit message from command line to a message file. In case a
whitelisted environment variable has a super long value and we run out
of characters on the command line.

Passing all Travis CI env variables starting with "TRAVIS" to Tox
testenv docsV (for commit message when updating docs in the future).