Fixing git.export() on Windows.
Piping to python (using tarfile built-in package) instead of the "tar"
command which is not installed on all Windows hosts.

Now that export() is handling untarring there is no need for a temporary
directory to extract to and copy from. Combined functionality into 1.5
iterations.

Fixes https://github.com/Robpol86/sphinxcontrib-versioning/issues/17